requirements.txt
# Core dependencies
streamlit==1.28.1
pandas==2.1.3
numpy==1.24.3
python-dotenv==1.0.0

# Data visualization
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.17.0
plotly-express==0.4.1

# AI/ML dependencies
openai==1.3.7
scikit-learn==1.3.2
httpx==0.27.2

# Image processing and PDF handling
pillow==10.0.1
PyMuPDF==1.23.8

# HTTP requests
requests==2.31.0

# Excel file handling
openpyxl==3.1.2
xlrd==2.0.1

# Streamlit extensions
streamlit-option-menu==0.3.6
streamlit-aggrid==0.3.4
streamlit-plotly-events==0.0.6

# Progress bars and utilities
tqdm==4.66.1
